refactor and adding utils
